requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
bleach==3.3.0
click==7.1.2
Django==3.1.7
django-bleach==0.6.1
djangorestframework==3.12.4
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
psycopg2-binary==2.8.6
PyAudio==0.2.11
pyparsing==2.4.7
pytz==2021.1
regex==2021.3.17
six==1.15.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
webencodings==0.5.1
